feat(storybook-config): update storybook configuration - remove ally test addon and unused config - Add .test.* paths to Storybook config (due to storybookTest plugin)
