+++ packages/ui/.storybook/preview.ts
@@ -1,0 +1,16 @@
+import type { Preview } from "@storybook/react"
+import "./index.css"
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+    layout: "centered",
+  },
+}
+
+export default preview
