+++ packages/ui/.storybook/vitest.setup.ts
@@ -1,13 +1,9 @@
-import * as a11yAddonAnnotations from "@storybook/addon-a11y/preview"
 import { beforeAll } from "vitest"
 import { setProjectAnnotations } from "@storybook/react"
 import * as projectAnnotations from "./preview"
 
 // This is an important step to apply the right configuration when testing your stories.
 // More info at: https://storybook.js.org/docs/api/portable-stories/portable-stories-vitest#setprojectannotations
-const project = setProjectAnnotations([
-  a11yAddonAnnotations,
-  projectAnnotations,
-])
+const project = setProjectAnnotations([projectAnnotations])
 
 beforeAll(project.beforeAll)
